+++ apps/next/util/stringUtils.ts
@@ -1,0 +1,3 @@
+export const formatUsd = (amount: number, negate?: boolean) => {
+    return `$${(amount / (100 * (negate ? -1 : 1))).toFixed(2)}`;
+};